+++ app/views/admin/citations/_table.html.erb
@@ -7,7 +7,7 @@
         <th>
           <%= sortable "created_at", skip_sortable: skip_sortable %>
         </th>
-        <% if render_updated_at %>
+        <% if render_updated_at? %>
           <th>
             <small><%= sortable "updated_at", skip_sortable: skip_sortable %></small>
           </th>
@@ -31,7 +31,7 @@
             <% if display_dev_info? %><code class="only-dev-visible-small"><%= citation.id %></code><% end %>
             <span class="convertTime"><%= l(citation.created_at, format: :convert_time) %></span>
           </td>
-          <% if render_updated_at %>
+          <% if render_updated_at? %>
             <td>
               <small class="convertTime"><%= l(citation.updated_at, format: :convert_time) %></small>
             </td>
