+++ app/views/admin/publishers/index.html.erb
@@ -10,7 +10,7 @@
         <th>
           <%= sortable "created_at", skip_sortable: skip_sortable %>
         </th>
-        <% if render_updated_at %>
+        <% if render_updated_at? %>
           <th>
             <small><%= sortable "updated_at", skip_sortable: skip_sortable %></small>
           </th>
@@ -37,7 +37,7 @@
               <% if display_dev_info? %><code class="only-dev-visible-small"><%= publisher.id %></code><% end %>
               <%= link_to l(publisher.created_at, format: :convert_time), edit_admin_publisher_path(publisher), class: "convertTime" %>
             </td>
-            <% if render_updated_at %>
+            <% if render_updated_at? %>
               <td>
                 <small class="convertTime"><%= l(publisher.updated_at, format: :convert_time) %></small>
               </td>
