+++ app/views/admin/topics/_table.html.erb
@@ -8,7 +8,7 @@
         <th>
           <%= sortable "created_at", skip_sortable: skip_sortable %>
         </th>
-        <% if render_updated_at %>
+        <% if render_updated_at? %>
           <th>
             <small><%= sortable "updated_at", skip_sortable: skip_sortable %></small>
           </th>
@@ -35,7 +35,7 @@
             <% if display_dev_info? %><code class="only-dev-visible-small"><%= topic.id %></code><% end %>
             <span class="convertTime"><%= l(topic.created_at, format: :convert_time) %></span>
           </td>
-          <% if render_updated_at %>
+          <% if render_updated_at? %>
             <td>
               <small class="convertTime"><%= l(topic.updated_at, format: :convert_time) %></small>
             </td>
