+++ spec/helpers/application_helper_spec.rb
@@ -3,109 +3,6 @@
 require "rails_helper"
 
 RSpec.describe ApplicationHelper, type: :helper do
-  describe "agreement_display" do
-    it "returns nil" do
-      expect(agreement_display("")).to be_nil
-    end
-    context "neutral" do
-      it "returns -" do
-        expect(agreement_display(:neutral)).to be_blank
-      end
-    end
-    context "agree" do
-      let(:target) { "<span title=\"Agree\"><img class=\"w-4 inline-block\" src=\"/images/icons/agree_icon.svg\" /></span>" }
-      it "returns -" do
-        expect(agreement_display(:agree)).to eq target
-      end
-      context "link" do
-        let(:target) { "<a title=\"Agree\" href=\"/ratings?search_agree=true&amp;search_disagree=false\"><img class=\"w-4 inline-block\" src=\"/images/icons/agree_icon.svg\" /></a>" }
-        let(:bp) { {controller: "ratings", action: "index"} }
-        it "returns with link" do
-          expect(agreement_display("agree", link: bp)).to eq target
-          expect(agreement_display(:agree, link: bp.merge(search_agree: false, search_disagree: true))).to eq target
-          expect(agreement_display("agree", link: bp.merge(search_disagree: false))).to eq target
-          expect(agreement_display(:agree, link: bp.merge(search_disagree: true, search_agree: false))).to eq target
-          # Same result if search_agreement == disagree
-          @search_agreement = :disagree
-          expect(agreement_display("agree", link: bp)).to eq target
-        end
-        # context "link: true" do
-        #   # TODO: need to stub current route I think? Not sure exactly what to do to make
-        #   # url_for() correctly pull the current controller_name and action_name in tests
-        #   it "returns with link" do
-        #     expect(agreement_display(:agree, link: true)).to eq target
-        #   end
-        # end
-        context "matching search_agreement" do
-          let(:target) { "<a title=\"Agree\" href=\"/ratings\"><img class=\"w-4 inline-block\" src=\"/images/icons/agree_icon.svg\" /></a>" }
-          before { @search_agreement = :agree }
-          it "returns with link with no agreement params" do
-            expect(agreement_display("agree", link: bp)).to eq target
-            expect(agreement_display(:agree, link: bp.merge(search_agree: true))).to eq target
-          end
-        end
-      end
-    end
-  end
-
-  describe "quality_display" do
-    it "returns nil" do
-      expect(quality_display("")).to be_nil
-    end
-    context "neutral" do
-      it "returns -" do
-        expect(quality_display("quality_med")).to be_nil
-      end
-    end
-    context "agree" do
-      let(:target) { "<span title=\"High Quality\"><img class=\"w-4 inline-block\" src=\"/images/icons/quality_high_icon.svg\" /></span>" }
-      it "returns -" do
-        expect(quality_display(:quality_high)).to eq target
-      end
-    end
-  end
-
-  describe "rating_display" do
-    let(:target) { "<span class=\" less-strong\">missing url</span>" }
-    it "returns target" do
-      rating = Rating.new
-      rating.display_name = rating.calculated_display_name
-      expect(rating_display(rating)).to eq target
-    end
-    context "with rating" do
-      let(:target) { "<a class=\" break-words\" title=\"#{citation.pretty_url}\" href=\"#{rating.submitted_url}\">texasattorneygeneral.gov/sites/default/files/images/admin/2021/Press/DC%20Statehood%20letter%20as%20sent%20(02539672x...</a>" }
-      let(:citation) { rating.citation }
-      let(:rating) { FactoryBot.create(:rating, submitted_url: "https://www.texasattorneygeneral.gov/sites/default/files/images/admin/2021/Press/DC%20Statehood%20letter%20as%20sent%20(02539672xD2C78)%20(002).pdf") }
-      it "returns target" do
-        expect(citation).to be_valid
-        expect(citation.title).to be_blank
-        expect(rating_display(rating)).to eq target
-      end
-    end
-    context "with rating with title" do
-      let(:target) { "<a class=\" break-words\" title=\"Somewhere\" href=\"https://example.com\">Somewhere</a>" }
-      let(:rating) { Rating.new(submitted_url: "https://example.com", citation_title: "Somewhere") }
-      it "returns target" do
-        rating.display_name = rating.calculated_display_name
-        expect(rating_display(rating)).to eq target
-      end
-    end
-  end
-
-  describe "citation_display" do
-    let(:target) { "<a class=\" break-words\" title=\"Somewhere\" href=\"https://example.com\">Somewhere</a>" }
-    let(:citation) { Citation.new(url: "https://example.com", title: "Somewhere") }
-    it "renders" do
-      expect(citation_display(citation)).to eq target
-    end
-    context "passed class" do
-      let(:target) { "<a class=\"otherClass break-words\" title=\"Somewhere\" href=\"https://example.com\">Somewhere</a>" }
-      it "renders" do
-        expect(citation_display(citation, {class: "otherClass"})).to eq target
-      end
-    end
-  end
-
   describe "page_description" do
     it "returns nil" do
       expect(page_description).to be_nil
